LayoutTests/imported/w3c/web-platform-tests/encoding/textdecoder-copy.any-expected.txt


FAIL Modify buffer after passing it in (ArrayBuffer) assert_equals: expected "@" but got "﻿@"
FAIL Modify buffer after passing it in (SharedArrayBuffer) undefined is not a constructor (evaluating 'new self[arrayBufferOrSharedArrayBuffer](2)')
